Extract github nwo helper functions
